+++ field/field-locator-contact.js
@@ -1,8 +1,8 @@
 
 const FieldLocator = require('./field-locator').FieldLocator;
 // const FieldTextBoolean = require('./field-text-boolean').FieldTextBoolean;
-const FieldText = require('./field-text').FieldText;
-const FieldGuid = require('./field-text').FieldTextGuid;
+const FieldLocatorText = require('./field-locator').FieldLocatorText;
+const FieldLocatorGuid = require('./field-locator').FieldLocatorGuid;
 const _ = require('lodash');
 
 class FieldLocatorContact extends FieldLocator {
@@ -11,14 +11,14 @@ class FieldLocatorContact extends FieldLocator {
     super(options);
     this._fields = _.merge(this._fields, {
       // _sync : new FieldTextBoolean(),
-      id: new FieldGuid({ emptyAllow: false}),
-      fullName: new FieldText({ emptyAllow: false}),
-      name: new FieldText(({emptyAllow: false})),
-      subName: new FieldText(({emptyAllow: false})),
-      guid: new FieldGuid({ emptyAllow: false}),
-      search: new FieldText({ emptyAllow: false}),
-      typeId: new FieldText({ emptyAllow: false}),
-      type: new FieldText({ emptyAllow: false}),
+      id: new FieldLocatorGuid({ emptyAllow: false}),
+      fullName: new FieldLocatorText({ emptyAllow: false}),
+      name: new FieldLocatorGuid({emptyAllow: false}),
+      subName: new FieldLocatorText({emptyAllow: false}),
+      guid: new FieldLocatorGuid({ emptyAllow: false}),
+      search: new FieldLocatorText({ emptyAllow: false}),
+      typeId: new FieldLocatorText({ emptyAllow: false}),
+      type: new FieldLocatorText({ emptyAllow: false}),
     });
   }
 }